Move commit actions
