+++ waimak-core/src/main/scala/com/coxautodata/waimak/dataflow/DataFlow.scala
@@ -216,81 +216,6 @@ trait DataFlow[Self <: DataFlow[Self]] extends Logging {
   }
 
   /**
-    * Groups labels to commit under a commit name.
-    * Can be called multiple times with same same commit name, thus adding labels to it.
-    * There can be multiple commit names defined in a single data flow.
-    *
-    * By default, the committer is requested to cache the underlying labels on the flow before writing them out
-    * if caching is supported by the data committer. If caching is not supported this parameter is ignored.
-    * This behavior can be disabled by setting the [[CACHE_REUSED_COMMITTED_LABELS]] parameter.
-    *
-    * @param commitName  name of the commit, which will be used to define its push implementation
-    * @param partitions  list of partition columns for the labels specified in this commit invocation. It will not
-    *                    impact labels from previous or following invocations of the commit with same commit name.
-    * @param repartition to repartition the data
-    * @param labels      labels added to the commit name with partitions config
-    * @return
-    */
-  def commit(commitName: String, partitions: Seq[String], repartition: Boolean = true)(labels: String*): Self = {
-    commit(commitName, Some(Left(partitions)), repartition = partitions.nonEmpty && repartition)(labels: _*)
-  }
-
-  /**
-    * Groups labels to commit under a commit name.
-    * Can be called multiple times with same same commit name, thus adding labels to it.
-    * There can be multiple commit names defined in a single data flow.
-    *
-    * By default, the committer is requested to cache the underlying labels on the flow before writing them out
-    * if caching is supported by the data committer. If caching is not supported this parameter is ignored.
-    * This behavior can be disabled by setting the [[CACHE_REUSED_COMMITTED_LABELS]] parameter.
-    *
-    * @param commitName  name of the commit, which will be used to define its push implementation
-    * @param repartition how many partitions to repartition the data by
-    * @param labels      labels added to the commit name with partitions config
-    * @return
-    */
-  def commit(commitName: String, repartition: Int)(labels: String*): Self = {
-    commit(commitName, Some(Right(repartition)), repartition = true)(labels: _*)
-  }
-
-  /**
-    * Groups labels to commit under a commit name.
-    * Can be called multiple times with same same commit name, thus adding labels to it.
-    * There can be multiple commit names defined in a single data flow.
-    *
-    * By default, the committer is requested to cache the underlying labels on the flow before writing them out
-    * if caching is supported by the data committer. If caching is not supported this parameter is ignored.
-    * This behavior can be disabled by setting the [[CACHE_REUSED_COMMITTED_LABELS]] parameter.
-    *
-    * @param commitName name of the commit, which will be used to define its push implementation
-    * @param labels     labels added to the commit name with partitions config
-    * @return
-    */
-  def commit(commitName: String)(labels: String*): Self = {
-    commit(commitName, None, repartition = false)(labels: _*)
-  }
-
-  private def commit(commitName: String, partitions: Option[Either[Seq[String], Int]], repartition: Boolean)(labels: String*): Self = {
-    val cacheLabels = this.flowContext.getBoolean(CACHE_REUSED_COMMITTED_LABELS, CACHE_REUSED_COMMITTED_LABELS_DEFAULT)
-
-    updateCommitMeta(_.addCommits(commitName, labels, partitions, repartition, cacheLabels))
-  }
-
-
-  /**
-    * Associates commit name with an implementation of a data committer. There must be only one data committer per one commit name.
-    *
-    * @param commitName
-    * @param committer
-    * @return
-    */
-  def push(commitName: String)(committer: DataCommitter[Self]): Self = updateCommitMeta(_.addPush(commitName, committer))
-
-  private def updateCommitMeta(update: CommitMeta[Self] => CommitMeta[Self]): Self =
-    updateExtensionMetadata(CommitExtension, m => update(m.getMetadataAsType[CommitMeta[Self]]))
-
-
-  /**
     * Creates a code block with all actions inside of it being run on the specified execution pool. Same execution pool
     * name can be used multiple times and nested pools are allowed, the name closest to the action will be assigned to it.
     *
